+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Requires `phpunit/php-code-coverage` `~14.2` to avoid using deprecated code
+- 
 ### Fixed
 
 - `$assert->debug()` was leaking the last successful shrunk values